Student and Lecture Controller have been completed

+++ src/main/java/com/demo/StudentRegistration/repository/StudentRepository.java
@@ -5,10 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository <Student,Long>  {
+public interface StudentRepository extends JpaRepository <Student,Integer>  {
     Student findById(int id);
     void deleteById (int id);
-
 }
